trying to use spdlog, but does not work as expected yet
